💄 rename: 관리자 container 폴더 구조 수정
