+++ aif_group/src/app/admin/manage/homepage/page.tsx
@@ -1,4 +1,4 @@
-import HomepageContent from '@/containers/admin/HomepageContent';
+import HomepageContent from '@/containers/admin/homepage/HomepageContent';
 import React from 'react';
 
 function HomePage() {
