+++ aif_group/src/app/admin/page.tsx
@@ -1,4 +1,4 @@
-import LoginForm from '@/containers/admin/LoginForm';
+import LoginForm from '@/containers/admin/login/LoginForm';
 import React from 'react';
 
 function LoginPage() {
