+++ aif_group/src/app/admin/signup/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SignupForm from '@/containers/admin/SignupFrom';
+import SignupForm from '@/containers/admin/login/SignupForm';
 
 function SignupPage() {
   return <SignupForm />;
