+++ aif_group/src/containers/admin/homepage/HomepageContent.tsx
@@ -1,35 +1,34 @@
 'use client';
+import { Switch } from 'antd';
 import React from 'react';
-import PromptModal from '../modal/PromptModal';
+import PromptModal from '../../modal/PromptModal';
 
-function UserContent() {
-  function handleClickPromptButton() {}
-  function handleClickFormButton() {}
-
+function HomepageContent() {
+  function handleClickPromptButton() {
+    // 사용자마다의 프롬프트 api를 받아오는 로직 ...
+  }
   return (
     <div className="mt-[3.3125rem] mx-[5.4375rem]">
       <table className="w-[100%] table-fixed text-[#828282] text-[1rem]">
         <thead className="h-[4.5625rem] bg-table_bg">
           <tr className="font-normal">
-            <th className="border-t-[1px] border-table_border w-[8.9490968801%]">번호</th>
-            <th className="border-t-[1px] border-x-[1px] border-table_border w-[20.68308703%]">아이디</th>
-            <th className="border-t-[1px] border-x-[1px] border-table_border w-[12.8078817734%]">사용일시</th>
-            <th className="border-t-[1px] border-x-[1px] border-table_border w-[10.0476190476%]">API 사용 횟수</th>
-            <th className="border-t-[1px] border-x-[1px] border-table_border w-[18%]">프롬프트/이미지 No</th>
-            <th className="border-t-[1px] border-table_border w-[18%]">설문 No</th>
+            <th className="border-t-[1px] border-table_border w-[5.9490968801%]">번호</th>
+            <th className="border-t-[1px] border-x-[1px] border-table_border w-[15.68308703%]">아이디(이메일)</th>
+            <th className="border-t-[1px] border-x-[1px] border-table_border w-[25.8078817734%]">프롬프트 / 이미지</th>
+            <th className="border-t-[1px] border-x-[1px] border-table_border w-[10.0476190476%]">생년일</th>
+            <th className="border-t-[1px] border-table_border w-[6.039408867%]">노출 활성화</th>
           </tr>
         </thead>
         <tbody>
           <tr className="h-[3.375rem]">
             <td className="bg-white border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
-            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
-            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border text-center">
-              <PromptModal onSelect={handleClickPromptButton}>adfas</PromptModal>
+              <PromptModal onSelect={handleClickPromptButton}>프롬프트1</PromptModal>
             </td>
+            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border text-center"></td>
             <td className="bg-white border-b-[1px] border-table_border text-center">
-              <button onClick={handleClickFormButton}>설문1</button>
+              <Switch />
             </td>
           </tr>
 
@@ -38,6 +37,13 @@ function UserContent() {
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
+            <td className="bg-white border-b-[1px] border-table_border"></td>
+          </tr>
+
+          <tr className="h-[3.375rem]">
+            <td className="bg-white border-b-[1px] border-table_border"></td>
+            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
+            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-b-[1px] border-table_border"></td>
           </tr>
@@ -47,29 +53,17 @@ function UserContent() {
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
-            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-b-[1px] border-table_border"></td>
           </tr>
 
           <tr className="h-[3.375rem]">
             <td className="bg-white border-b-[1px] border-table_border"></td>
-            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
-            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
-            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
-            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
-            <td className="bg-white border-b-[1px] border-table_border"></td>
-          </tr>
-
-          <tr className="h-[3.375rem]">
-            <td className="bg-white border-b-[1px] border-table_border"></td>
-            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-b-[1px] border-table_border"></td>
           </tr>
           <tr className="h-[4.5625rem]">
-            <td className="bg-table_bg border-b-[1px] border-table_border"></td>
             <td className="bg-table_bg border-b-[1px] border-table_border"></td>
             <td className="bg-table_bg border-b-[1px] border-table_border"></td>
             <td className="bg-table_bg border-b-[1px] border-table_border"></td>
@@ -82,4 +76,4 @@ function UserContent() {
   );
 }
 
-export default UserContent;
+export default HomepageContent;
