+++ aif_group/src/containers/admin/user/UserContent.tsx
@@ -1,34 +1,35 @@
 'use client';
-import { Switch } from 'antd';
 import React from 'react';
-import PromptModal from '../modal/PromptModal';
+import PromptModal from '../../modal/PromptModal';
 
-function HomepageContent() {
-  function handleClickPromptButton() {
-    // 사용자마다의 프롬프트 api를 받아오는 로직 ...
-  }
+function UserContent() {
+  function handleClickPromptButton() {}
+  function handleClickFormButton() {}
+
   return (
     <div className="mt-[3.3125rem] mx-[5.4375rem]">
       <table className="w-[100%] table-fixed text-[#828282] text-[1rem]">
         <thead className="h-[4.5625rem] bg-table_bg">
           <tr className="font-normal">
-            <th className="border-t-[1px] border-table_border w-[5.9490968801%]">번호</th>
-            <th className="border-t-[1px] border-x-[1px] border-table_border w-[15.68308703%]">아이디(이메일)</th>
-            <th className="border-t-[1px] border-x-[1px] border-table_border w-[25.8078817734%]">프롬프트 / 이미지</th>
-            <th className="border-t-[1px] border-x-[1px] border-table_border w-[10.0476190476%]">생년일</th>
-            <th className="border-t-[1px] border-table_border w-[6.039408867%]">노출 활성화</th>
+            <th className="border-t-[1px] border-table_border w-[8.9490968801%]">번호</th>
+            <th className="border-t-[1px] border-x-[1px] border-table_border w-[20.68308703%]">아이디</th>
+            <th className="border-t-[1px] border-x-[1px] border-table_border w-[12.8078817734%]">사용일시</th>
+            <th className="border-t-[1px] border-x-[1px] border-table_border w-[10.0476190476%]">API 사용 횟수</th>
+            <th className="border-t-[1px] border-x-[1px] border-table_border w-[18%]">프롬프트/이미지 No</th>
+            <th className="border-t-[1px] border-table_border w-[18%]">설문 No</th>
           </tr>
         </thead>
         <tbody>
           <tr className="h-[3.375rem]">
             <td className="bg-white border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
+            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
+            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border text-center">
-              <PromptModal onSelect={handleClickPromptButton}>프롬프트1</PromptModal>
+              <PromptModal onSelect={handleClickPromptButton}>adfas</PromptModal>
             </td>
-            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border text-center"></td>
             <td className="bg-white border-b-[1px] border-table_border text-center">
-              <Switch />
+              <button onClick={handleClickFormButton}>설문1</button>
             </td>
           </tr>
 
@@ -37,13 +38,6 @@ function HomepageContent() {
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
-            <td className="bg-white border-b-[1px] border-table_border"></td>
-          </tr>
-
-          <tr className="h-[3.375rem]">
-            <td className="bg-white border-b-[1px] border-table_border"></td>
-            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
-            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-b-[1px] border-table_border"></td>
           </tr>
@@ -53,17 +47,29 @@ function HomepageContent() {
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
+            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-b-[1px] border-table_border"></td>
           </tr>
 
           <tr className="h-[3.375rem]">
             <td className="bg-white border-b-[1px] border-table_border"></td>
+            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
+            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
+            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
+            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
+            <td className="bg-white border-b-[1px] border-table_border"></td>
+          </tr>
+
+          <tr className="h-[3.375rem]">
+            <td className="bg-white border-b-[1px] border-table_border"></td>
+            <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-x-[1px] border-b-[1px] border-table_border"></td>
             <td className="bg-white border-b-[1px] border-table_border"></td>
           </tr>
           <tr className="h-[4.5625rem]">
+            <td className="bg-table_bg border-b-[1px] border-table_border"></td>
             <td className="bg-table_bg border-b-[1px] border-table_border"></td>
             <td className="bg-table_bg border-b-[1px] border-table_border"></td>
             <td className="bg-table_bg border-b-[1px] border-table_border"></td>
@@ -76,4 +82,4 @@ function HomepageContent() {
   );
 }
 
-export default HomepageContent;
+export default UserContent;
